Reorder tests
