+++ colonists-shared/__tests__/harbors.spec.ts
@@ -1,0 +1,41 @@
+/*
+* Harbors
+*/
+describe('Rules when trading using harbors', () => {
+    
+    test('A player trading with a 3:1 harbor with no house on a 3:1 harbor tile is not allowed', () => {
+  
+    });
+  
+    test('A player trading with a 3:1 harbor with a house on a 3:1 harbor tile is allowed', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor with no house on a 2:1 harbor tile is not allowed', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor with a house on a 2:1 harbor tile is allowed', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor on a harbor with a different resource type is not allowed', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor on a harbor with the corresponding resource type is allowed', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor on a harbor with the corresponding resource type decreases the amount of the given resource by 2', () => {
+  
+    });
+  
+    test('A player trading with a 2:1 harbor on a harbor with the corresponding resource type increases the amount of the received resource by 1', () => {
+  
+    });
+  
+    test('A player trading with a 3:1 harbor on a harbor increases the amount of the received resource by 1', () => {
+  
+    });
+  });
