+++ colonists-shared/__tests__/trading.spec.ts
@@ -1,0 +1,29 @@
+/* 
+* Trading
+*/
+describe('Rules when trading with other players', () => {
+
+    test('A player trading with another player will lose the resource traded away', () => {
+  
+    });
+  
+    test('A player trading with another player for will gain the resource traded for', () => {
+  
+    });
+  
+    test('A player trading 4:1 with the bank with 4 resources of the same type is allowed', () => {
+  
+    });
+  
+    test('A player trading 4:1 with the bank with 4 resources of different types is not allowed', () => {
+  
+    });
+  
+    test('A player trading 4:1 with the bank with 4 resources of the same types loses 4 resources of that type', () => {
+  
+    });
+  
+    test('A player trading 4:1 with the bank with 4 resources of the same types gains 1 resources of the chosen type', () => {
+  
+    });
+  });
